Changes to build & deploy lwraft UI independently on any host.
Also inclued are the RPM build/installation changes to lightwaveui

+++ ui/lightwave-ui/build/lightwave-ui.spec
@@ -32,6 +32,7 @@ npm run prod
 %install
 install -dm755 %{buildroot}
 install -dm755 %{buildroot}/opt/vmware/lightwaveui
+install -dm755 %{buildroot}/opt/vmware/config
 install -dm755 %{buildroot}/opt/vmware/lightwaveui/directory
 mkdir -p %{buildroot}/opt/vmware/tools
 
@@ -40,23 +41,46 @@ cp -a dist/* %{buildroot}/opt/vmware/lightwaveui/
 find %{buildroot}/opt/vmware/lightwaveui -type d -exec chmod 755 {} \;
 find %{buildroot}/opt/vmware/lightwaveui -type f -exec chmod 644 {} \;
 
-install -m644 ../build/nginx.conf %{buildroot}/opt/vmware/lightwaveui/config
-install -m755 ../build/oidc-client-utils %{buildroot}/opt/vmware/tools
+install -m644 ../build/nginx_lightwaveui.conf %{buildroot}/opt/vmware/config
+install -m755 ../build/oidc-client-utils-lightwaveui %{buildroot}/opt/vmware/tools
 
 cd ../lwdir
 install -m644 dist/* %{buildroot}/opt/vmware/lightwaveui/directory
 
 %post
-mv /etc/nginx/nginx.conf /etc/nginx/nginx_orig.conf
-ln -s /opt/vmware/lightwaveui/config/nginx.conf /etc/nginx/nginx.conf
+lwraftui_cfg="/opt/vmware/config/nginx_lwraftui.conf"
+lightwaveuiui_cfg="/opt/vmware/config/nginx_lightwaveui.conf"
+if [ -f "$lwraftui_cfg" ]
+then
+    cp $lwraft_cfg $lightwaveui_cfg
+    sed -i '$ d' $lightwaveui_cfg
+    sed -i '$ d' $lightwaveui_cfg
+    echo "
+        location /lightwaveui {
+            root /opt/vmware;
+            index  index.html index.htm;
+        }
+    }
+}" >> $lightwaveui_cfg
+    rm /etc/nginx/nginx.conf
+else
+    mv /etc/nginx/nginx.conf /etc/nginx/nginx_orig.conf
+fi
+ln -s /opt/vmware/config/nginx_lightwaveui.conf /etc/nginx/nginx.conf
 systemctl restart nginx
-echo "Please register this lightwave ui instance as OIDC client on lightwave server as shown below:"
-echo "/opt/vmware/tools/oidc-client-utils"
-/opt/vmware/tools/oidc-client-utils register -e
+echo "Please register this lightwave UI instance as OIDC client on lightwave server as shown below:"
+echo "/opt/vmware/tools/oidc-client-utils-lightwaveui"
+/opt/vmware/tools/oidc-client-utils-lightwaveui register -e
 
 %postun
 rm /etc/nginx/nginx.conf
-mv /etc/nginx/nginx_orig.conf /etc/nginx/nginx.conf
+lwraftui_cfg="/opt/vmware/config/nginx_lwraftui.conf"
+if [ -f "$lwraftui_cfg" ]
+then
+    ln -s $lwraftui_cfg /etc/nginx/nginx.conf
+else
+    mv /etc/nginx/nginx_orig.conf /etc/nginx/nginx.conf
+fi
 systemctl restart nginx
 
 %files
@@ -64,6 +88,7 @@ systemctl restart nginx
 %dir /opt/vmware/lightwaveui
 /opt/vmware/lightwaveui
 /opt/vmware/tools
+/opt/vmware/config
 
 %changelog
 *    Mon Jan 08 2018 Harish Udaiya Kumar <hudaiyakumar@vmware.com> 0.1-1
